Fixed the serialization

+++ Playlist/PlaylistControl.cs
@@ -1,30 +1,37 @@
-﻿using System;
+﻿using Byte_me___Group_2.Playlist;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Diagnostics;
 using System.Drawing;
+using System.Drawing.Text;
 using System.IO;
 using System.Linq;
+using System.Runtime.Remoting.Metadata.W3cXsd2001;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-//ACTUALLLLL
 namespace Byte_me___Group_2
 {
     public partial class PlaylistControl : UserControl
     {
         private Home homeForm;
 
-
         private string username, coversFolder, playlistsFolder, dataFolder;
         OpenFileDialog ofdCoverPicture;
         Panel pnlMainContent, pnlSidebar;
         Button btnNewPlaylist, btnChangeCoverPhoto;
 
+        // The list of songs in the playlist
         private BindingList<Song> Songs;
+
+        // The class for the song that is currently playing
         CurrentlyPlaying currentlyPlayingSong;
+
+        SongHistory SongsQueue;
+
 
         public PlaylistControl(Home homeForm, string username, string coversFolder, string playListsFolder, string dataFolder, OpenFileDialog ofdCoverPicture, Panel pnlMain, Panel Sidebar, Button btnNewPlayList, Button btnChangeCover)
         {
@@ -40,14 +47,15 @@ namespace Byte_me___Group_2
             this.btnChangeCoverPhoto = btnChangeCover;
             this.pnlSidebar = Sidebar;
             this.homeForm = homeForm;
+
+
+            playerControls1.setSongPlayer(wmpSongPlay);
         }
 
-        ////Lists for the songNames, artists, and durations
-        //List<string> songs = new List<string>();
-        //List<string> artists = new List<string>();
-        //List<string> songDurations = new List<string>();
-
-        //Basic display setup when the user opens a playlist
+        /// <summary>
+        /// Basic display setup when the user opens a playlist
+        /// </summary>
+        /// <param name="title"> The playlist that the user opened </param>
         public void setupPlaylistPage(string title)
         {
             lblPlaylistName.Text = title;
@@ -55,6 +63,7 @@ namespace Byte_me___Group_2
             lblDateCreated.Text = null;
             lblWelecome.Text = $"Welcome back, {this.username}";
             
+            // Read the songs from the textfile
             readSongs(title);
             dgvDisplaySongs.DataSource = Songs;
 
@@ -66,8 +75,11 @@ namespace Byte_me___Group_2
 
             lblPlalistCount.Text = $"This playlist has {Songs.Count} songs";
 
+            //Gets the creation date of the playlist and displays it
             string creationDate = File.GetCreationTime(Path.Combine(playlistsFolder, $"{title}.txt")).ToString("dd MMMM yyyy");
             addToCreationLabel(creationDate);
+
+            SongsQueue = new SongHistory( playlistsFolder, title );
         }
 
         //Goes back to the home screen when the user clicks on the filepath
@@ -217,6 +229,7 @@ namespace Byte_me___Group_2
             }
         }
 
+        //
         private void dgvDisplaySongs_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             // Gaurd against the header being clicked
@@ -226,17 +239,28 @@ namespace Byte_me___Group_2
             }
 
             // Deletes the song which delete button is pressed
-            if (dgvDisplaySongs.Columns[e.ColumnIndex].Name == "SongDelete")
+            if (e.ColumnIndex > 0 && dgvDisplaySongs.Columns[e.ColumnIndex].Name == "SongDelete")
             {
                 DeleteSong(e.RowIndex, lblPlaylistName.Text);
             } 
             else
             {
+                //Plays the song that is clicked
                 wmpSongPlay.URL = Songs[e.RowIndex].SongFilePath;
                 wmpSongPlay.Ctlcontrols.play();
 
                 currentlyPlayingSong = new CurrentlyPlaying();
+
+                playSong(e.RowIndex);
             }
+        }
+
+        public void playSong(int rowIndex)
+        {
+            playerControls1.lblTime.Text = dgvDisplaySongs[2, rowIndex].Value.ToString();
+
+            playerControls1.tmrTrackbarTime.Enabled = true;
+            
         }
 
        
@@ -302,12 +326,9 @@ namespace Byte_me___Group_2
             {
                 foreach (Song song in Songs)
                 {
-                    writer.WriteLine($"{song.Name}|{song.Artist}|{song.Duration}");
+                    writer.WriteLine($"{song.Name}|{song.Artist}|{song.Duration}|{song.SongFilePath}");
                 }
             }
-
-            // Rerenders all the remaining songs to display
-            //readSongs(playlistName);
 
             // Shows a message for successful deletion
             MessageBox.Show("Song deleted successfully.");
